Module Api

module Parameters : sig ... end

Contains module types for defining parameters of API endpoints, and modules that implement these types when relevant

module Responses : sig ... end

Contains module types for defining responses of API endpoints, and modules that implement these types when relevant.

module Specification : sig ... end