Module Ocoi

OCaml On Ice library

module Controllers : sig ... end

Provides functionality for defining Ice controllers and registering them with an Opium app.

module Db : sig ... end

Various functions involving database access.

module Handler_utils : sig ... end

Utility functions for handling requests to an Opium app.