Module Auth.Checks

Defines various check functions to be used by authenticate.

val accept_all : auth_credential option -> Opium.Std.Request.t -> bool

Allows all requests

val reject_all : auth_credential option -> Opium.Std.Request.t -> bool

Rejects all requests as unauthorized

val bearer_only : (string -> Opium.Std.Request.t -> bool) -> auth_credential option -> Opium.Std.Request.t -> bool

bearer_only token_check rejects all requests with an Authorization header not of the form Authorization: Bearer <token>, and uses token_check to determine whether requests that are of that form should be allowed

val jwt : algorithm:Jwt.algorithm -> validate:(Jwt.payload -> Opium.Std.Request.t -> bool) -> auth_credential option -> Opium.Std.Request.t -> bool

jwt_check ~algorithm ~validate produces an auth check that uses JWT tokens

It rejects requests with a JWT token that cannot be parsed or has an incorrect signature according to algorithm, and validates other tokens against the request using validate.