Module Ocoi.Authentication

Implements authentication for requests on a per route basis using the Authorization HTTP header

Example use:

let my_route =
  get "/my_route"
    (Ocoi.Authentication.Checks.accept_all (fun _ ->
         `String "Hello world!" |> respond'))

implements a very simple kind of "authentication" that accepts all requests to /my_route.

A more useful example would replace Ocoi.Authentication.Checks.accept_all with some other function with signature auth_credential option -> Request.t -> bool that takes an Authorization head and returns whether it is valid for the associated request. For instance, using JWT tokens for authentication:

let jwt_algorithm = Jwt.HS256 "SupaSekretKey"

let check_user jwt_payload req =
  let jwt_user = Jwt.(find_claim (claim "user_id") jwt_payload) in
  let req_id = param req "id" in
  jwt_user = req_id

let get_user =
  get "/users/:id"
    (Ocoi.Authentication.Checks.jwt_check ~algorithm:jwt_algorithm ~validate:check_user (fun _ ->
         get_user (param req "id") |> respond'))

type auth_credential = [

| Cohttp.Auth.credential

| `Bearer of string

]

Extends Cohttp authorization header types with Bearer

val authenticate : check:(auth_credential option -> Opium.Std.Request.t -> bool) -> (Opium.Std.Request.t -> Opium.Std.Response.t Lwt.t) -> Opium.Std.Request.t -> Opium.Std.Response.t Lwt.t

authenticate ~check handler_body wraps handler_body by using check for authentication

When called on a request, the new function will parse the Authorization header of the request and call check on that and the request. If this gives true the original handler body will be called. Otherwise an empty 401 Unauthorized response will be returned.

module Checks : sig ... end

Defines various check functions to be used by authenticate.