Module Ocoi__.Ocoi_middleware

val allow_cors : ?⁠origin:string -> ?⁠methods:string Core.sexp_list -> unit -> Opium.Std.Rock.Middleware.t

allow_cors origin methods enables Cross-Origin Resource Sharing for requests from origin with methods in the list methods.

If origin is not given it defaults to all origins (the wildcard value "*"). If methods is not given it defaults to all methods.