Module Ocoi__Controllers

Provides functionality for defining Ice controllers and registering them with an Opium app

module type Crud = sig ... end

Represents Create, Read, Update and Delete operations on some model t

val create_handler : string -> (module Crud) -> Opium.Std.App.builder

app |> create_handler "name" (module Crud) exposes Crud.create at POST /name

val index_handler : string -> (module Crud) -> Opium.Std.App.builder

app |> index_handler "name" (module Crud) exposes Crud.index at GET /name

val show_handler : string -> (module Crud) -> Opium.Std.App.builder

app |> show_handler "name" (module Crud) exposes Crud.show at GET /name/:id

val update_handler : string -> (module Crud) -> Opium.Std.App.builder

app |> update_handler "name" (module Crud) exposes Crud.update at PUT /name

Note that this differs from the conventional location of PUT /name/:id for the Update operation of CRUD. This is because the ID of the new model will always be supplied in the body, so doesn't need to be put in the URL too.

val destroy_handler : string -> (module Crud) -> Opium.Std.App.builder

app |> destroy_handler "name" (module Crud) exposes Crud.destroy at Delete /name/:id

val register_crud : string -> (module Crud) -> Opium.Std.App.t -> Opium.Std.App.t

register_crud "name" (module Crud) app calls index_handler, show_handler, update_handler and destroy_handler on app.