Module Ocoi_api

include Specification

type verb =

| Get

| Post

| Put

| Delete

val verb_to_route : verb -> Opium.Std.App.route

module type S = sig ... end

module type Implementation = sig ... end

module Parameters : sig ... end

Contains module types for defining parameters of API endpoints, and modules that implement these types when relevant

module Responses : sig ... end

Contains module types for defining responses of API endpoints, and modules that implement these types when relevant.

module Specification : sig ... end