Module Ocoi_api.Parameters

Contains module types for defining parameters of API endpoints, and modules that implement these types when relevant

module type Json = sig ... end

For endpoints that take a piece of JSON

module type Jwt_json = sig ... end

For endpoints that take a piece of JSON and a JWT token

module type None = sig ... end

For endpoints that take no parameters

module None : sig ... end

module type One_param = sig ... end

For endpoints with a single parameter in the path of the URL, for example id in example.com/things/:id

module One_param : sig ... end

Contains implementations of One_param for common types

module type Custom = sig ... end

For custom parameter specifications that supply some function f for producing the parameter type from an Opium request