Module Path.One_and

For endpoints with a single path parameter and some other non-path parameters

module type Query = sig ... end

For endpoints with a single path parameter and some query parameters