Module type Parameters.One_param

For endpoints with a single parameter in the path of the URL, for example id in example.com/things/:id

Not for endpoints with a single query parameters (query parameters are not yet implemented). The general case of multiple path parameters is also not yet implemented, but will be.

type t

val of_string : string -> t