Module Ocoi_api.Responses

Contains module types for defining responses of API endpoints, and modules that implement these types when relevant.

For more details, see Ocoi_handlers__.Responses.Make.

type status_code

Represents an HTTP status code

module Implementations : sig ... end

Defines simple modules and module types that used as the basis for other response specification modules/types.

module type Json = sig ... end

For endpoints that return a piece of JSON

module type No_content = Implementations.Unit

For endpoints that return an empty response with a 204 No content code

module No_content = Implementations.Unit

module Created : sig ... end

For endpoints that return an empty response and a Location header with a URL with a single path parameter

module type Empty_code = sig ... end

module Empty_code : sig ... end

module type Empty_code_headers = sig ... end

module Empty_code_headers : sig ... end

module type String = sig ... end

module String : sig ... end

module type Json_list = Json

Json_list, Json_opt and Json_code are aliases for Json, * and the functors with these names don't do anything. But they should still be used for endpoints will be used with * Ocoi_handlers__.Responses.Make.Json_list and similar for documentation purposes and to enable (currently hypothetical)

module type Json_opt = Json

module type Json_code = Json

module Json_list : functor (Some_json : Json) -> Json with type t = Some_json.t

module Json_opt : functor (Some_json : Json) -> Json with type t = Some_json.t

module Json_code : functor (Some_json : Json) -> Json with type t = Some_json.t