Module type Responses.Json_list

Json_list, Json_opt and Json_code are aliases for Json, * and the functors with these names don't do anything. But they should still be used for endpoints will be used with * Ocoi_handlers__.Responses.Make.Json_list and similar for documentation purposes and to enable (currently hypothetical)

type t

val yojson_of_t : t -> Yojson.Safe.t