Module Ocoi_api__.Parameters

module type Json = sig ... end

For endpoints that take a piece of JSON

module type Jwt_json = sig ... end

For endpoints that take a piece of JSON and a JWT token

module type None = sig ... end

For endpoints that take no parameters

module None : sig ... end

module Path : sig ... end

For endpoints parameters in the path of the URL, for example id in example.com/things/:id

module type Jwt_path_one = sig ... end

module Jwt_path_one_int : sig ... end

module type Custom = sig ... end

For custom parameter specifications that supply some function f for producing the parameter type from an Opium request

module type Jwt = sig ... end

module Jwt : Jwt