Module Ocoi_api__Parameters.Path

For endpoints parameters in the path of the URL, for example id in example.com/things/:id

module type One = sig ... end

For endpoints where the only parameter is a single path parameter. Multiple path parameters will be implemented in the future.

module One : sig ... end

Contains implementations of One for common types

module One_and : sig ... end

For endpoints with a single path parameter and some other non-path parameters