OCaml On Ice

OCaml On Ice is a web framework in the style of Ruby on Rails, built on top of Opium. It is designed for building REST APIs, especially those that are consumed by BuckleScript or js_of_ocaml frontends.

There is a guide to using it in the GitHub README and reference documentation here Ocoi.