Module Ocoi

The OCaml On Ice library

module App : sig ... end

Defines an alternative to Opium.Std.App.empty that already has some middleware applied

module Authentication : sig ... end

Implements authentication for requests on a per route basis using the Authorization HTTP header

module Controllers : sig ... end

Provides functionality for defining Ice controllers and registering them with an Opium app

module Db : sig ... end

Defines various functions involving database access

module Handler_utils : sig ... end

Utility functions for handling requests to an Opium app

module Jwt_utils : sig ... end

Defines utilities for working with JWT tokens

module Logging : sig ... end

Provides logging functionality for Ice apps. Logging is done using the Logs library as in Opium.

module Middleware = Ocoi__.Ocoi_middleware

Defines Opium middlewares for use in Ice apps