OCaml On Ice

OCaml On Ice is a web framework in the style of Ruby on Rails, built on top of Opium. It is designed for building REST APIs, especially those that are consumed by BuckleScript or js_of_ocaml frontends.

Getting started

The best way to get started with OCaml On Ice is following the Tutorial (in the README) and reading the design philosophy page and glossary.

Documentation

Commands documentation

Library reference

Design philosophy

Glossary

Project structure reference

Roadmap

Known issues

Other links

GitHub